refactor(battery)!: Remove deprecated battery code
This removes the deprecated code that searches for a device labeled
"BATTERY" to use as the battery state of charge sensor. Battery
reporting now requires using the zmk,battery chosen node.

To integrate this change with a custom board, search for any node with
a "label" property set to "BATTERY":

    vbatt {
        label = "BATTERY";
        ...
    };

If you find one, remove the "label" property. Add a node label if it is
missing one:

    vbatt: vbatt {
        ...
    };

Then, find the "chosen" node and set the "zmk,battery" property to
reference the node:

    chosen {
        ...
        zmk,battery = &vbatt;
        ...
    };